+++ CPAC/image_utils/__init__.py
@@ -1,2 +1,22 @@
-from .spatial_smoothing import *
-from .statistical_transforms import *
+# Copyright (C) 2018-2022  C-PAC Developers
+
+# This file is part of C-PAC.
+
+# C-PAC is free software: you can redistribute it and/or modify it under
+# the terms of the GNU Lesser General Public License as published by the
+# Free Software Foundation, either version 3 of the License, or (at your
+# option) any later version.
+
+# C-PAC is distributed in the hope that it will be useful, but WITHOUT
+# ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or
+# FITNESS FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public
+# License for more details.
+
+# You should have received a copy of the GNU Lesser General Public
+# License along with C-PAC. If not, see <https://www.gnu.org/licenses/>.
+from .spatial_smoothing import set_gauss, spatial_smoothing
+from .statistical_transforms import calc_avg, fisher_z_score_standardize, \
+                                    z_score_standardize
+
+__all__ = ['calc_avg', 'fisher_z_score_standardize', 'set_gauss',
+           'spatial_smoothing', 'z_score_standardize']
