+++ CPAC/image_utils/statistical_transforms.py
@@ -1,7 +1,22 @@
-from CPAC.pipeline import nipype_pipeline_engine as pe
-import nipype.interfaces.utility as util
-from nipype.interfaces.afni import preprocess
+# Copyright (C) 2018-2022  C-PAC Developers
 
+# This file is part of C-PAC.
+
+# C-PAC is free software: you can redistribute it and/or modify it under
+# the terms of the GNU Lesser General Public License as published by the
+# Free Software Foundation, either version 3 of the License, or (at your
+# option) any later version.
+
+# C-PAC is distributed in the hope that it will be useful, but WITHOUT
+# ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or
+# FITNESS FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public
+# License for more details.
+
+# You should have received a copy of the GNU Lesser General Public
+# License along with C-PAC. If not, see <https://www.gnu.org/licenses/>.
+from nipype.interfaces import utility as util
+from nipype.interfaces.afni import preprocess
+from CPAC.pipeline import nipype_pipeline_engine as pe
 from CPAC.utils import function
 from CPAC.utils.utils import (
     extract_output_mean,
